[ci skip] Merge pull request #254 from sparcians/knutel/issue_253_extension_ignored
Changed the behavior of Sparta's extension support -- it's now an error if the extension does not exist in the tree OR is not referenced by the simulator. Added support for <OPTIONAL> on the extensions.

Fixed an issue where -c <config> and --arch yaml files would pick up extension changes, but not -p in the same order as regular parameters. e778f2f0da06601e9a3f22ec1c0ef6e31902e39b